Put both all-teams and specific-team pages under 1 teams folder, only allow user to view team data if registered in comp (hardcoded)
