--- src/pages/Competitions/CompetitionAllTeamsPage/index.less
@@ -1,4 +1,0 @@
-@import '../../../styles/base.less';
-.CompetitionAllTeamsPage {
-  
-}
