--- src/pages/Competitions/CompetitionSpecificTeamPage/index.less
@@ -1,3 +1,0 @@
-@import '../../../styles/base.less';
-.CompetitionSpecificTeamPage {
-}
