+++ src/pages/Competitions/CompetitionTeamPages/AllTeamsPage/index.less
@@ -1,0 +1,4 @@
+@import '../../../../styles/base.less';
+.CompetitionAllTeamsPage {
+  
+}
