+++ src/pages/Competitions/CompetitionTeamPages/SpecificTeamPage/index.less
@@ -1,0 +1,3 @@
+@import '../../../../styles/base.less';
+.CompetitionSpecificTeamPage {
+}
